feat(sync): warn on config refs dangled by rename
A rename (explicit or collision auto-rename) can leave config-side name
references pointing at agent/skill names that no longer exist, with no
diagnostic. Surface a `config-rename-dangle` warning for each:

- [settings.meridian.fanout].agents entries
- [agents.<name>] overlay keys (config + local)
- [skills.<name>] overlay keys

One `warn_config_dangles_after_rename` covers all three, driven by both
rename sources against the post-prune installed set, so a silently
mismatched config no longer goes undiagnosed.

Extracted validation helpers to `src/sync/validate.rs` and moved the
inline test module to `src/sync/tests.rs` (following the discover/tests.rs
pattern), bringing mod.rs from 2494 to 868 lines.
